refactor: drop latexdiff wrapper
